[SBL-44] 리팩토링 진행
