+++ src/app/(main)/dashboard/page.module.css
@@ -1,4 +1,4 @@
-.sectionsWrapper {
+.fieldsWrapper {
   width: 100%;
   padding: 12px 0 48px;
