+++ src/app/(main)/layout.tsx
@@ -1,7 +1,7 @@
 import { Header, Footer } from '@/components/layout/main';
 import styles from './layout.module.css';
 
-function MainLayout({ children }: Readonly<{ children: React.ReactNode }>) {
+function Layout({ children }: Readonly<{ children: React.ReactNode }>) {
   return (
     <div className={styles.layout}>
       <Header />
@@ -11,4 +11,4 @@ function MainLayout({ children }: Readonly<{ children: React.ReactNode }>) {
   );
 }
 
-export default MainLayout;
+export default Layout;
