+++ src/app/(main)/page.tsx
@@ -1,11 +1,12 @@
 import Banner from '@/components/main/banner/Banner';
 import { SurveyFinder } from '@/components/main';
+import { Suspense } from 'react';
 
-export default function Home() {
+export default function Page() {
   return (
-    <>
+    <Suspense>
       <Banner />
       <SurveyFinder />
-    </>
+    </Suspense>
   );
 }
