+++ src/components/survey-p/funcs/index.ts
