+++ src/components/survey-p/types/interaction.ts
