+++ src/components/survey-p/ui/navigator/Navigator.module.css
