+++ src/components/survey-p/ui/navigator/Navigator.tsx
