+++ src/providers/ReactQueryClientProvider.tsx
@@ -3,7 +3,7 @@
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { useState } from 'react';
 
-export function ReactQueryClientProvider({ children }: { children: React.ReactNode }) {
+export default function ReactQueryClientProvider({ children }: { children: React.ReactNode }) {
   const [queryClient] = useState(
     () =>
       new QueryClient({
