+++ src/services/utils.ts
@@ -1,0 +1,10 @@
+const API_URL = process.env.NEXT_PUBLIC_API_URL;
+
+const makeUrl = (api: string[], params?: { [key: string]: string }) => {
+  const url = new URL([API_URL, ...api].join('/'));
+  if (params) Object.keys(params).forEach((key) => url.searchParams.append(key, params[key]));
+
+  return url.toString();
+};
+
+export { makeUrl };
