--- src/utils/index.ts
@@ -1,4 +1,0 @@
-import { pushURLSearchParams, replaceURLSearchParams } from './urlSearchParams';
-import { dateReader } from './misc';
-
-export { pushURLSearchParams, replaceURLSearchParams, dateReader };
